MCQ constructeur + documentation SQLiteHelper

+++ MiniPoll/app/src/main/java/be/lsinf1225gr12/minipoll/minipoll/MySQLiteHelper.java
@@ -51,7 +51,7 @@ ContentValues cv = new ContentValues();
 String selection = NOM_COLONNE + " = ? AND " + AUTRE_NOM_COLONNE + " = ?"; //rajouter autant qu'il faut
 String[] selectionArgs = new String[]{String.valueOf(VALEUR), String.valueOf(VALEUR)};
 cv.update(NIM_TABLE, cv, selection, selectionArgs);
-int result = db.insert(NOM_TABLE, null, cv);
+int result = (int) db.insert(NOM_TABLE, null, cv);
 if (result==-1)
 {
 	//erreur dans l'ajout, suppression
